+++ naga/src/proc/mod.rs
@@ -11,7 +11,7 @@ mod terminator;
 mod typifier;
 
 pub use constant_evaluator::{
-    ConstantEvaluator, ConstantEvaluatorError, ExpressionConstnessTracker,
+    ConstantEvaluator, ConstantEvaluatorError, ExpressionKind, ExpressionKindTracker,
 };
 pub use emitter::Emitter;
 pub use index::{BoundsCheckPolicies, BoundsCheckPolicy, IndexableLength, IndexableLengthError};
@@ -216,8 +216,8 @@ impl crate::Literal {
             (value, crate::ScalarKind::Sint, 4) => Some(Self::I32(value as _)),
             (value, crate::ScalarKind::Uint, 8) => Some(Self::U64(value as _)),
             (value, crate::ScalarKind::Sint, 8) => Some(Self::I64(value as _)),
-            (1, crate::ScalarKind::Bool, 4) => Some(Self::Bool(true)),
-            (0, crate::ScalarKind::Bool, 4) => Some(Self::Bool(false)),
+            (1, crate::ScalarKind::Bool, crate::BOOL_WIDTH) => Some(Self::Bool(true)),
+            (0, crate::ScalarKind::Bool, crate::BOOL_WIDTH) => Some(Self::Bool(false)),
             _ => None,
         }
     }
@@ -553,13 +553,9 @@ impl crate::Expression {
     ///
     /// [`Access`]: crate::Expression::Access
     /// [`ResolveContext`]: crate::proc::ResolveContext
-    pub fn is_dynamic_index(&self, module: &crate::Module) -> bool {
+    pub const fn is_dynamic_index(&self) -> bool {
         match *self {
-            Self::Literal(_) | Self::ZeroValue(_) => false,
-            Self::Constant(handle) => {
-                let constant = &module.constants[handle];
-                !matches!(constant.r#override, crate::Override::None)
-            }
+            Self::Literal(_) | Self::ZeroValue(_) | Self::Constant(_) => false,
             _ => true,
         }
     }
@@ -652,7 +648,8 @@ impl crate::Module {
         GlobalCtx {
             types: &self.types,
             constants: &self.constants,
-            const_expressions: &self.const_expressions,
+            overrides: &self.overrides,
+            global_expressions: &self.global_expressions,
         }
     }
 }
@@ -667,17 +664,18 @@ pub(super) enum U32EvalError {
 pub struct GlobalCtx<'a> {
     pub types: &'a crate::UniqueArena<crate::Type>,
     pub constants: &'a crate::Arena<crate::Constant>,
-    pub const_expressions: &'a crate::Arena<crate::Expression>,
+    pub overrides: &'a crate::Arena<crate::Override>,
+    pub global_expressions: &'a crate::Arena<crate::Expression>,
 }
 
 impl GlobalCtx<'_> {
-    /// Try to evaluate the expression in `self.const_expressions` using its `handle` and return it as a `u32`.
+    /// Try to evaluate the expression in `self.global_expressions` using its `handle` and return it as a `u32`.
     #[allow(dead_code)]
     pub(super) fn eval_expr_to_u32(
         &self,
         handle: crate::Handle<crate::Expression>,
     ) -> Result<u32, U32EvalError> {
-        self.eval_expr_to_u32_from(handle, self.const_expressions)
+        self.eval_expr_to_u32_from(handle, self.global_expressions)
     }
 
     /// Try to evaluate the expression in the `arena` using its `handle` and return it as a `u32`.
@@ -700,7 +698,7 @@ impl GlobalCtx<'_> {
         &self,
         handle: crate::Handle<crate::Expression>,
     ) -> Option<crate::Literal> {
-        self.eval_expr_to_literal_from(handle, self.const_expressions)
+        self.eval_expr_to_literal_from(handle, self.global_expressions)
     }
 
     fn eval_expr_to_literal_from(
@@ -724,7 +722,7 @@ impl GlobalCtx<'_> {
         }
         match arena[handle] {
             crate::Expression::Constant(c) => {
-                get(*self, self.constants[c].init, self.const_expressions)
+                get(*self, self.constants[c].init, self.global_expressions)
             }
             _ => get(*self, handle, arena),
         }